Double quote mess cleaned up!
