+++ lib/src/testing_library/dom/queries/by_placeholder_text.dart
@@ -9,7 +9,8 @@ import 'package:js/js.dart';
 import 'package:over_react_test/src/testing_library/dom/async/types.dart';
 import 'package:over_react_test/src/testing_library/dom/matches/types.dart';
 import 'package:over_react_test/src/testing_library/dom/queries/interface.dart';
-import 'package:over_react_test/src/testing_library/util/error_message_utils.dart' show promiseToFutureWithErrorInterop, withErrorInterop;
+import 'package:over_react_test/src/testing_library/util/error_message_utils.dart'
+    show promiseToFutureWithErrorInterop, withErrorInterop;
 
 /// PRIVATE. Do not export from this library.
 ///
@@ -143,7 +144,7 @@ mixin ByPlaceholderTextQueries on IQueries {
     NormalizerFn Function(NormalizerOptions) normalizer,
     Duration timeout,
     Duration interval,
-    /*Error*/dynamic Function(/*Error*/dynamic originalError) onTimeout,
+    /*Error*/ dynamic Function(/*Error*/ dynamic originalError) onTimeout,
     MutationObserverOptions mutationObserverOptions = defaultMutationObserverOptions,
   }) {
     final matcherOptions = buildMatcherOptions(exact: exact, normalizer: normalizer);
@@ -185,7 +186,7 @@ mixin ByPlaceholderTextQueries on IQueries {
     NormalizerFn Function(NormalizerOptions) normalizer,
     Duration timeout,
     Duration interval,
-    /*Error*/dynamic Function(/*Error*/dynamic originalError) onTimeout,
+    /*Error*/ dynamic Function(/*Error*/ dynamic originalError) onTimeout,
     MutationObserverOptions mutationObserverOptions = defaultMutationObserverOptions,
   }) {
     final matcherOptions = buildMatcherOptions(exact: exact, normalizer: normalizer);
