+++ test/testing_library/util/constants.dart
@@ -1,0 +1,4 @@
+import 'package:test/test.dart' show Timeout;
+
+const asyncQueryTimeout = Duration(milliseconds: 200);
+final asyncQueryTestTimeout = Timeout(asyncQueryTimeout * 2);
